アノテーションを使用した、編集ウィンドウ自動生成を実装 選択式プロパティの実装

+++ src/interpolation/Interpolation.kt
@@ -4,5 +4,5 @@ package interpolation
  * キーフレームの補完をする抽象クラス
  */
 abstract class Interpolation {
-    abstract fun getInterpolation(input : Float):Float
+    abstract fun getInterpolation(input : Double):Double
 }
