+++ src/interpolation/LinearInterpolation.kt
@@ -4,7 +4,7 @@ package interpolation
  * 直線補間を行う
  */
 class LinearInterpolation : Interpolation() {
-    override fun getInterpolation(input: Float):Float {
+    override fun getInterpolation(input: Double):Double  {
         return input
     }
 }
